# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brainiac",
    "description": "GitOps for organizational AI knowledge — capture, govern, and serve an organization's memory. Every read is RLS-scoped to the caller: an agent can never retrieve what its operator cannot.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8600",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/provision": {
      "post": {
        "tags": [
          "provision"
        ],
        "summary": "POST /v1/provision — exchange a verified identity for its project.",
        "description": "Exchange a VERIFIED federated identity for its single project, optionally minting a device key. Requires the `admin` scope; the caller must have already verified the identity with the provider.",
        "operationId": "provision_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project provisioned or already existed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported provider, or empty subject/email"
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          }
        }
      }
    },
...
    "/v1/tokens": {
      "get": {
        "tags": [
          "tokens"
        ],
        "summary": "GET /v1/tokens — list the org's tokens (metadata only, never secrets).",
        "description": "List the org's API tokens (metadata only; secrets are never retrievable).",
        "operationId": "list_tokens",
        "responses": {
          "200": {
            "description": "Tokens of the caller's org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          }
        }
      },
      "post": {
        "tags": [
          "tokens"
        ],
        "summary": "POST /v1/tokens — mint a token. The secret appears ONCE in this response;\nonly its sha256 is stored. Requires the `admin` scope (env bootstrap\ntokens qualify), so read/write tokens cannot mint tokens.",
        "description": "Mint an API token; the plaintext secret is returned once and never again.",
        "operationId": "create_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Token minted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "Empty name or scopes outside the mintable set (see auth::SCOPES)"
          },
          "401": {
            "description": "Missing or unknown bearer token"
          },
          "403": {
            "description": "Token lacks the `admin` scope"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTokenBody": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Scope the key to one project (migration 0034); omit for org-wide."
          },
          "scopes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Subset of read|write|admin; defaults to [\"read\"]."
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Principal the token acts as; defaults to the caller."
          }
        }
      },
...
      "CreatedTokenResponse": {
        "type": "object",
        "description": "The mint response. `token` is the plaintext secret and is the only place\nit will ever appear — everything else is retrievable from `GET /v1/tokens`.",
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "user_id",
          "token"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The project the key is scoped to; null = org-wide."
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "token": {
            "type": "string",
            "description": "Shown exactly once — never retrievable again."
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ProvisionBody": {
        "type": "object",
        "description": "A verified federated identity, plus what to do with it.",
        "required": [
          "provider",
          "subject",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "For display and support only; never the identity."
          },
          "issue_key": {
            "type": "boolean",
            "description": "Mint a fresh device key. Always true in effect on first creation (a project\nwith no key cannot be reached); on a returning identity it is opt-in, so a\npage refresh does not quietly mint keys."
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "What to call the project. Defaults to the email's local part."
          },
          "provider": {
            "type": "string",
            "description": "Identity provider. Only `google` today."
          },
          "subject": {
            "type": "string",
            "description": "The provider's stable subject — the Firebase uid. NOT the email: emails get\nreassigned and change case, uids do not."
          }
        }
      },
      "ProvisionResponse": {
        "type": "object",
        "required": [
          "org_id",
          "user_id",
          "team_id",
          "created"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "The device key, in the ONLY response that will ever contain it — only its\nsha256 is stored. Null when no key was issued this call."
          },
          "api_key_prefix": {
            "type": [
              "string",
              "null"
            ],
            "description": "The key's display prefix, safe to persist and show in a list."
          },
          "created": {
            "type": "boolean",
            "description": "False when the identity already had a project. Signing in twice is not an\nerror — it just does not make a second project."
          },
          "org_id": {
            "type": "string",
            "format": "uuid"
          },
          "team_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "TokenListResponse": {
        "type": "object",
        "required": [
          "tokens"
        ],
        "properties": {
          "tokens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TokenSummary"
            }
          }
        }
      },
...
      "TokenSummary": {
        "type": "object",
        "description": "One token as the listing exposes it — metadata only, never the secret.",
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string"
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The project the key is scoped to; null = org-wide."
          },
          "project_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name of that project; null = org-wide."
          },
          "revoked_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Env-map token (unrestricted) or a managed `brk_…` key carrying the read/write/admin scope the route requires."
      }
    }
  },
...
}